implement code reivew comments

+++ boot-high-rps-sample/src/main/java/com/example/highrps/batchprocessor/PostCommentBatchProcessor.java
@@ -146,6 +146,9 @@ public class PostCommentBatchProcessor implements EntityBatchProcessor {
                     .filter(Objects::nonNull)
                     .toList();
 
+            if (cacheKeysList.isEmpty()) {
+                return;
+            }
             try {
                 long deletedRows = postCommentRepository.deleteByCommentRefIdIn(cacheKeysList);
                 log.debug("Deleted : {} comment entity for ids: {}", deletedRows, cacheKeysList);
